Add currencies fragment
